+++ pkg/server/server.go
@@ -5,7 +5,6 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"slices"
 	"sort"
 	"time"
 
@@ -31,6 +30,7 @@ import (
 	"github.com/openfga/openfga/internal/utils/apimethod"
 	"github.com/openfga/openfga/pkg/authclaims"
 	"github.com/openfga/openfga/pkg/encoder"
+	"github.com/openfga/openfga/pkg/featureflags"
 	"github.com/openfga/openfga/pkg/gateway"
 	"github.com/openfga/openfga/pkg/logger"
 	serverconfig "github.com/openfga/openfga/pkg/server/config"
@@ -41,16 +41,11 @@ import (
 	"github.com/openfga/openfga/pkg/typesystem"
 )
 
-type ExperimentalFeatureFlag string
-
 const (
 	AuthorizationModelIDHeader = "Openfga-Authorization-Model-Id"
 	authorizationModelIDKey    = "authorization_model_id"
 
-	ExperimentalCheckOptimizations       ExperimentalFeatureFlag = "enable-check-optimizations"
-	ExperimentalListObjectsOptimizations ExperimentalFeatureFlag = "enable-list-objects-optimizations"
-	ExperimentalAccessControlParams      ExperimentalFeatureFlag = "enable-access-control"
-	allowedLabel                                                 = "allowed"
+	allowedLabel = "allowed"
 )
 
 var tracer = otel.Tracer("openfga/pkg/server")
@@ -173,10 +168,11 @@ type Server struct {
 	maxAuthorizationModelCacheSize   int
 	maxTypesystemCacheSize           int
 	maxAuthorizationModelSizeInBytes int
-	experimentals                    []ExperimentalFeatureFlag
+	experimentals                    []string
 	AccessControl                    serverconfig.AccessControlConfig
 	AuthnMethod                      string
 	serviceName                      string
+	featureFlagClient                featureflags.Client
 
 	// NOTE don't use this directly, use function resolveTypesystem. See https://github.com/openfga/openfga/issues/1527
 	typesystemResolver     typesystem.TypesystemResolverFunc
@@ -400,9 +396,20 @@ func WithMaxConcurrentReadsForListUsers(maxConcurrentReadsForListUsers uint32) O
 	}
 }
 
-func WithExperimentals(experimentals ...ExperimentalFeatureFlag) OpenFGAServiceV1Option {
+func WithExperimentals(experimentals ...string) OpenFGAServiceV1Option {
 	return func(s *Server) {
 		s.experimentals = experimentals
+	}
+}
+
+func WithFeatureFlagClient(client featureflags.Client) OpenFGAServiceV1Option {
+	return func(s *Server) {
+		if client != nil {
+			s.featureFlagClient = client
+			return
+		}
+
+		s.featureFlagClient = featureflags.NewNoopFeatureFlagClient()
 	}
 }
 
@@ -595,13 +602,10 @@ func MustNewServerWithOpts(opts ...OpenFGAServiceV1Option) *Server {
 	return s
 }
 
-func (s *Server) IsExperimentallyEnabled(flag ExperimentalFeatureFlag) bool {
-	return slices.Contains(s.experimentals, flag)
-}
-
 // IsAccessControlEnabled returns true if the access control feature is enabled.
 func (s *Server) IsAccessControlEnabled() bool {
-	return s.IsExperimentallyEnabled(ExperimentalAccessControlParams) && s.AccessControl.Enabled
+	isEnabled := s.featureFlagClient.Boolean(serverconfig.ExperimentalAccessControlParams, nil)
+	return isEnabled && s.AccessControl.Enabled
 }
 
 // WithListObjectsDispatchThrottlingEnabled sets whether dispatch throttling is enabled for List Objects requests.
@@ -838,7 +842,7 @@ func NewServerWithOpts(opts ...OpenFGAServiceV1Option) (*Server, error) {
 		maxAuthorizationModelSizeInBytes: serverconfig.DefaultMaxAuthorizationModelSizeInBytes,
 		maxAuthorizationModelCacheSize:   serverconfig.DefaultMaxAuthorizationModelCacheSize,
 		maxTypesystemCacheSize:           serverconfig.DefaultMaxTypesystemCacheSize,
-		experimentals:                    make([]ExperimentalFeatureFlag, 0, 10),
+		experimentals:                    make([]string, 0, 10),
 		AccessControl:                    serverconfig.AccessControlConfig{Enabled: false, StoreID: "", ModelID: ""},
 
 		cacheSettings: serverconfig.NewDefaultCacheSettings(),
@@ -914,6 +918,10 @@ func NewServerWithOpts(opts ...OpenFGAServiceV1Option) (*Server, error) {
 
 	if s.listUsersDispatchThrottlingMaxThreshold != 0 && s.listUsersDispatchDefaultThreshold > s.listUsersDispatchThrottlingMaxThreshold {
 		return nil, fmt.Errorf("ListUsers default dispatch throttling threshold must be equal or smaller than max dispatch threshold for ListUsers")
+	}
+
+	if s.featureFlagClient == nil {
+		s.featureFlagClient = featureflags.NewDefaultClient(s.experimentals)
 	}
 
 	err := s.validateAccessControlEnabled()